refactor walletdb driver tests into walletdbtest pkg and fix for api changes.
