+++ walletdb/walletdbtest/doc.go
@@ -1,0 +1,8 @@
+// Copyright (c) 2017 The btcsuite developers
+// Use of this source code is governed by an ISC
+// license that can be found in the LICENSE file.
+
+// Package walletdbtest provides exported tests that can be imported and
+// consumed by walletdb driver tests to help ensure that drivers confirm to the
+// database driver interface correctly.
+package walletdbtest
